Move StorageWiggler into its own module
